improve "get_technology_from_item" method and separate utils from shared
